move MockRequest class
