+++ tests/utils.py
@@ -48,3 +48,23 @@ def select_test_db(measure):
 
     result = requests.get(url, params=params)
     return result.json()
+
+
+class MockRequest(object):
+    """
+        Create mock request object for test.
+        borrow from influxdb-python.
+    """
+    def __init__(self, retries):
+        self.i = 0
+        self.retries = retries
+
+    def connection_error(self, *args, **kwargs):
+        self.i += 1
+
+        if self.i < self.retries:
+            raise requests.exceptions.ConnectionError
+        else:
+            r = requests.Response()
+            r.status_code = 204
+        return r
